extend pg support to all of the `1.x` major version
`pg` 1.2 was released on [2019-12-20](https://github.com/ged/ruby-pg/blob/v1.2.0/History.rdoc#v120-2019-12-20-michael-granger-gedfaeriemudorg).

Rails has had a version constraint on `pg` of `>= 0.18, < 2.0` since Rails 5.0, so Active Record's compatibility with the `pg` gem should not be an issue:

* [6.0](https://github.com/rails/rails/blob/v6.0.2.1/railties/lib/rails/generators/database.rb#L17)
* [5.2](https://github.com/rails/rails/blob/v5.2.4.1/railties/lib/rails/generators/app_base.rb#L304)
* [5.1](https://github.com/rails/rails/blob/v5.1.7/railties/lib/rails/generators/app_base.rb#L276)
* [5.0](https://github.com/rails/rails/blob/v5.0.7.2/railties/lib/rails/generators/app_base.rb#L272)

And as for the `pg` gem itself, it follows Semantic Versioning, so using a pessimistic version constraint should be safe (it's suggested in the [README](https://github.com/ged/ruby-pg#versioning)).