build: add ci workflow which replicates circleci config (#28)
This is close to a 1-1 replica of functionality.

No upgrades to build dependencies have been made, even when no longer supported upstream. Those are tagged for follow-up (`FU`) in separate PRs.

Changes:
- Removed code climate coverage reports as we don't have the integration setup

## Why are we doing this?

Circle no longer has the appropriate permissions to build this repository. Given that we're moving to Github Actions, it seems sensible to make the switch instead of fixing it.